fixed trySpawnBlock so it makes the active piece before checking if null
Changed the condition in Loader function so it is col <  4 not 5

Stored the possible kinds in an array

Created a new helper method in BlockyGame which shuffles an array and then used that method to shuffle the array that holds the kind.

Use the first element in the array as the kind of block.

Added proccessMovement() to step
Added break in the right movement