fixed trySpawnBlock so it makes the active piece before checking if null
Changed the condition in Loader function so it is col <  4 not 5

Stored the possible kinds in an array

Created a new helper method in BlockyGame which shuffles an array and then used that method to shuffle the array that holds the kind.

Use the first element in the array as the kind of block.

Added proccessMovement() to step
Added break in the right movement

Changed to completeRows.add(rows) in getCompletedRows() so that it returns an int of the row instead of the bool

Change the condition in deleteRow(int n), so that it clears the row first
and then for the for loop that shifts the blocks down, row is set the the row input of the function and is decremented until 0 which will shift all the blocks above down one.